requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
dateparser==1.1.8
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
numpy==1.26.1
pandas==2.1.2
pycryptodome==3.19.0
python-binance==1.0.19
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
regex==2023.10.3
requests==2.31.0
six==1.16.0
ta==0.11.0
tzdata==2023.3
tzlocal==5.2
ujson==5.8.0
urllib3==2.0.7
websockets==12.0
yarl==1.9.2
matplotlib==3.8.1
